requirements.txt
hubspot3
azureml-sdk[automl]
pandas
scikit-learn # If needed for preprocessing
